Move test utils.ts one folder up
