+++ packages/dom/test/text-quote/describe.test.ts
@@ -24,7 +24,7 @@ import { describeTextQuote } from '../../src/text-quote/describe';
 
 import testCases from './describe-cases';
 import testMatchCases from './match-cases';
-import { hydrateRange, evaluateXPath } from './utils';
+import { hydrateRange, evaluateXPath } from '../utils';
 
 const domParser = new window.DOMParser();
