+++ packages/dom/test/text-quote/match.test.ts
@@ -25,7 +25,7 @@ import { createTextQuoteSelectorMatcher } from '../../src/text-quote/match';
 import { DomScope } from '../../src/types';
 
 import testCases from './match-cases';
-import { evaluateXPath, RangeInfo } from './utils';
+import { evaluateXPath, RangeInfo } from '../utils';
 
 const domParser = new window.DOMParser();
